TASK: Introduce user.dn setting in favor of bind.dn
Before this change the bind.dn was used for binding, but also
for authenticating a user. The latter requires a placeholder ?
in the setting that will lead to bind errors if the LDAP server
requires a bind, as the ? is an invalid filter.

This change separates the two so the dn for the user authentication
can defer from the dn used for binding.

To make the change non-breaking the bind provider falls back to the
'old' setting if available.