TASK: Replace bind providers by symfony/ldap
The configuration of the bind providers lead to a lot of confusion
so we replace those by the symfony/ldap component. This change
is the first kickstart to a working setup with simplified configuration
format.

It also changes the way a user is authenticated from using a single
filter to directly fetch an account to a query that first fetches
a possible account and then uses the fetched 'dn' to authenticate
the user.

Additionaly a provider first pragmatic Neos backend provider is added.